requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.5.0
python-dotenv==1.0.0
pydantic==2.4.2
pydantic[email]==2.4.2